refactor utils and types to adhere to code structure
