+++ apps/backend-cli/src/actions/database/import-steady.ts
@@ -368,4 +368,3 @@ export const importSteady = async (
     console.log('Import completed successfully');
   });
 };
-
